renamed folder + simplified tests renamed the folder test to tests simplified the tests for the purpose of figuring out how to run them in docker
